refactor(xstate-inspector): bundle the UI and move state to zustand
The single browser/app.js is split by purpose under browser/src, bundled with
esbuild into browser/dist, and its state now lives in a zustand store that the
timeline subscribes to, so no action renders by hand any more.

The middleware serves browser/dist, which is generated and ignored, and
main-app:start builds it, so a development start still needs nothing extra.
Editing the UI needs a rebuild, or the new watch target.

@io-app/xstate-inspector becomes a devDependency of the app, matching the
bridge, which already returns undefined outside development bundles. esbuild is
listed in pnpm's allowBuilds because pnpm 11 fails an install while a dependency
build script stays skipped, which would have broken installs for everyone.
