+++ apps/main-app/package.json
@@ -3,7 +3,7 @@
   "version": "3.39.0-rc.2",
   "private": true,
   "scripts": {
-    "start": "react-native start",
+    "start": "nx run xstate-inspector:build && react-native start",
     "sync": "pnpm install && bundle && pnpm dev-pod-install && pnpm generate",
     "run-ios": "react-native run-ios",
     "run-android": "react-native run-android",
@@ -48,7 +48,6 @@
     "@gorhom/bottom-sheet": "^5.2.8",
     "@io-app/api-types": "workspace:*",
     "@io-app/design-system": "workspace:@pagopa/io-app-design-system@*",
-    "@io-app/xstate-inspector": "workspace:*",
     "@pagopa/io-pagopa-commons": "^3.1.0",
     "@pagopa/io-react-native-cie": "1.3.6",
     "@pagopa/io-react-native-cieid": "0.5.2",
@@ -173,6 +172,7 @@
     "@eslint/compat": "^2.0.3",
     "@eslint/eslintrc": "^3.3.5",
     "@eslint/js": "^9.39.4",
+    "@io-app/xstate-inspector": "workspace:*",
     "@pagopa/eslint-config": "6.2.2",
     "@pagopa/openapi-codegen-ts": "^14.2.0",
     "@react-native-community/cli": "20.0.0",
