+++ libs/xstate-inspector/package.json
@@ -7,6 +7,8 @@
   "types": "src/index.ts",
   "react-native": "src/index.ts",
   "scripts": {
+    "build": "node browser/build.mjs",
+    "watch": "node browser/build.mjs --watch",
     "test": "jest --config='./jest.config.js'",
     "tsc-noemit": "tsc --noEmit",
     "lint": "eslint \"**/*.ts\""
@@ -15,8 +17,10 @@
     "@react-native/typescript-config": "catalog:",
     "@types/jest": "catalog:",
     "@types/node": "^22.20.1",
+    "esbuild": "^0.28.2",
     "jest": "catalog:",
-    "typescript": "catalog:"
+    "typescript": "catalog:",
+    "zustand": "^5.0.15"
   },
   "dependencies": {
     "@statelyai/inspect": "0.7.2",
