+++ libs/xstate-inspector/tsconfig.json
@@ -5,5 +5,5 @@
     "types": ["jest", "node"],
     "lib": ["es2022", "dom", "dom.iterable"]
   },
-  "include": ["src", "middleware.js", "browser/app.js", "browser/types.d.ts"]
+  "include": ["src", "middleware.js", "browser/src", "browser/build.mjs"]
 }
